feat: split service account token fetching behavior
