+++ frontend/graphql/queries/service-accounts/getServiceAccountDetail.gql
@@ -19,18 +19,6 @@ query GetServiceAccountDetail($orgId: ID!, $id: ID) {
         self
       }
     }
-    tokens {
-      id
-      name
-      createdAt
-      expiresAt
-      createdBy {
-        fullName
-        avatarUrl
-        self
-      }
-      lastUsed
-    }
     appMemberships {
       id
       name
