+++ frontend/graphql/queries/service-accounts/getServiceAccountTokens.gql
@@ -1,0 +1,17 @@
+query GetServiceAccountTokens($orgId: ID!, $id: ID) {
+  serviceAccounts(orgId: $orgId, serviceAccountId: $id) {
+    id
+    tokens {
+      id
+      name
+      createdAt
+      expiresAt
+      createdBy {
+        fullName
+        avatarUrl
+        self
+      }
+      lastUsed
+    }
+  }
+}
